// src/enums/ErrorCode.ts
export enum ErrorCode {
  // Metadata validation
  MISSING_PROJECT_ID = 'MISSING_PROJECT_ID',
  INVALID_PROJECT_ID = 'INVALID_PROJECT_ID',
  MISSING_SERVICE_ID = 'MISSING_SERVICE_ID',
  INVALID_SERVICE_ID = 'INVALID_SERVICE_ID',
  MISSING_VERSION = 'MISSING_VERSION',
  INVALID_VERSION = 'INVALID_VERSION',

  // ActionsCoreLib config
  MISSING_ActionsCoreLib_TYPE = 'MISSING_ActionsCoreLib_TYPE',
  INVALID_ActionsCoreLib_TYPE = 'INVALID_ActionsCoreLib_TYPE',
  CONFIG_PARSE_ERROR = 'CONFIG_PARSE_ERROR',
  CONFIG_NOT_FOUND = 'CONFIG_NOT_FOUND',

  // Stage validation
  MISSING_REQUIRED_STAGE = 'MISSING_REQUIRED_STAGE',
  INVALID_STAGE_ORDER = 'INVALID_STAGE_ORDER',
  MISSING_STAGE_COMMANDS = 'MISSING_STAGE_COMMANDS',
  INVALID_STAGE_TYPE = 'INVALID_STAGE_TYPE',
  UNKNOWN_STAGE = 'UNKNOWN_STAGE',

  // Plan / deploy correspondence
  NO_PLAN_COMMAND_FOUND = 'NO_PLAN_COMMAND_FOUND',
  PLAN_COMMAND_FORBIDDEN = 'PLAN_COMMAND_FORBIDDEN',
  DEPLOY_PLAN_COMMAND_FORBIDDEN = 'DEPLOY_PLAN_COMMAND_FORBIDDEN',
  DEPLOY_MISSING_PLAN_REF = 'DEPLOY_MISSING_PLAN_REF',

  // Deploy validation
  MISSING_DEPLOY_ENV = 'MISSING_DEPLOY_ENV',
  INVALID_DEPLOY_ENV = 'INVALID_DEPLOY_ENV',
  ACCOUNT_BRANCH_MISMATCH = 'ACCOUNT_BRANCH_MISMATCH',
  APPROVER_NOT_AUTHORIZED = 'APPROVER_NOT_AUTHORIZED',
  INVALID_CONFIRM = 'INVALID_CONFIRM',

  // Artifact / publish
  MISSING_IMAGE_REF = 'MISSING_IMAGE_REF',
  MISSING_REGISTRY_CONFIG = 'MISSING_REGISTRY_CONFIG',

  // Hotfix emergency
  HOTFIX_EXCEPTION_NOT_ALLOWED = 'HOTFIX_EXCEPTION_NOT_ALLOWED',

  // Analyzer
  SONARQUBE_QUALITY_GATE_FAILED = 'SONARQUBE_QUALITY_GATE_FAILED',
  TRIVY_VULNERABILITIES_FOUND = 'TRIVY_VULNERABILITIES_FOUND',
  CHECKOV_POLICY_FAILED = 'CHECKOV_POLICY_FAILED',


  // Generic
  EXECUTION_FAILED = 'EXECUTION_FAILED',
}